+++ R/pro.R
@@ -65,8 +65,6 @@
 #' In other words: Non-Standard Evaluation. \cr
 #' \cr
 #' Regular Standard Evaluation does not have the above problem. \cr
-#' Standard evaluation in 'R' is not limited to atomic objects like character vectors;
-#' formulas can also be used.
 #' 
 #' 
 #'