+++ R/pversion.R
@@ -78,8 +78,7 @@ pversion_check4mismatch <- function(pkgs = NULL, lib.loc = .libPaths()) {
   
   if(length(pkgs) > 0) {
     versions_loaded <- lapply(pkgs, getNamespaceVersion) |> unlist()
-    versions_lib <- lapply(pkgs, \(x)as.character(utils::packageVersion(x, lib.loc = lib.loc))) |>
-      unlist()
+    versions_lib <- vapply(pkgs, \(x) .pversion_installed(x, lib.loc), character(1L))
     versions_compare <- mapply(utils::compareVersion, versions_loaded, versions_lib)
     ind <- which(versions_compare != 0)
     if(length(ind) > 0) {
@@ -121,8 +120,7 @@ pversion_report <- function(pkgs = NULL, lib.loc = .libPaths()) {
   
   if(length(pkgs) > 0) {
     versions_loaded <- lapply(pkgs, getNamespaceVersion) |> unlist()
-    versions_lib <- lapply(pkgs, \(x)as.character(utils::packageVersion(x, lib.loc = lib.loc))) |>
-      unlist()
+    versions_lib <- vapply(pkgs, \(x) .pversion_installed(x, lib.loc), character(1L))
     versions_compare <- mapply(utils::compareVersion, versions_loaded, versions_lib)
     versions_compare <- as.logical(versions_compare == 0)
     tab <- data.frame(
@@ -146,4 +144,19 @@ pversion_report <- function(pkgs = NULL, lib.loc = .libPaths()) {
 }
 
 
-
+#' @keywords internal
+#' @noRd
+.pversion_installed <- function(pkg, lib.loc) {
+  out <- character(1L)
+  for(i in lib.loc) {
+    temp <- tryCatch(
+      read.dcf(file.path(i, pkg, "DESCRIPTION"), fields = "Version"),
+      error = function(x) return(NA)
+    )
+    if(!is.na(temp)) {
+      out <- temp
+    }
+  }
+  
+  return(out)
+}